secured the orders endpoint

+++ server/models/orders.js
@@ -8,7 +8,7 @@ const { pool } = require('./db');
  */
 const getOrders = (request, response) => {
   var query = 'SELECT * FROM "Orders" inner join "Person" on "Person"."ID"="Orders"."ReceiverPersonID"';
-  query = authenticate(request,response, query);
+  query = authenticate(request, response, query);
 
   pool.query(query, (error, results) => {
     if (error) {
@@ -17,7 +17,6 @@ const getOrders = (request, response) => {
     response.status(200).json(results.rows);
   });
 };
-
 
 
 /**
@@ -30,7 +29,7 @@ const getOrdersDetails = (request, response) => {
 
   var query = 'SELECT * FROM "Orders" inner join "Address" on "Orders"."DropAddressID"="Address"."AddressID"';
   query = authenticate(request, response, query);
-  query +=  'ORDER BY "OrderID" ASC';
+  query += 'ORDER BY "OrderID" ASC';
 
   pool.query(query, (error, results) => {
     if (error) {
@@ -47,16 +46,14 @@ const getOrdersDetails = (request, response) => {
  * @param {obj} response response object * 
  */
 const getOrderById = (request, response) => {
-  
+
   const id = parseInt(request.params.id, 10);
 
-  var query ='SELECT 	"Orders".*,	"addrop"."StreetAddress" as dropstreetAddress, 	"addrop"."City" as dropcity,'+
-	'"addrop"."Country" as dropcountry, "addrop"."PostCode" as droppostcode, 	"adpick"."StreetAddress" as pickstreetAddress,'+
-  '"adpick"."City" as pickcity,	"adpick"."Country" as pickcountry,	"adpick"."PostCode" as pickpostcode FROM "Orders" inner join "Address" addrop on "Orders"."DropAddressID"="addrop"."AddressID" '+ 
-  'inner join "Address" adpick on "Orders"."PickAddressID"="adpick"."AddressID" ';
-
-  authenticate(request,response, query);
-
+  var query = 'SELECT 	"Orders".*,	"addrop"."StreetAddress" as dropstreetAddress, 	"addrop"."City" as dropcity,' +
+    '"addrop"."Country" as dropcountry, "addrop"."PostCode" as droppostcode, 	"adpick"."StreetAddress" as pickstreetAddress,' +
+    '"adpick"."City" as pickcity,	"adpick"."Country" as pickcountry,	"adpick"."PostCode" as pickpostcode FROM "Orders" inner join "Address" addrop on "Orders"."DropAddressID"="addrop"."AddressID" ' +
+    'inner join "Address" adpick on "Orders"."PickAddressID"="adpick"."AddressID" ';
+  authenticate(request, response, query);
   query += 'and "Orders"."OrderID" = $1'
 
   pool.query(query, [id], (error, results) => {
@@ -74,7 +71,7 @@ const getOrderById = (request, response) => {
  * @param {obj} response response object * 
  */
 
- //TODO: hardcoded for dummy implementation
+//TODO: hardcoded for dummy implementation
 const getPackageTimeline = (request, response) => {
   // const id = parseInt(request.params.pkgid);
 
@@ -106,21 +103,21 @@ const getPackageTimeline = (request, response) => {
  * @param {obj} response response object * 
  */
 const createOrder = (request, response) => {
-  if(request.userType == 2 || request.userType == 1){
-  const {
-    pickaddressid, dropaddressid, pickdate, arrivaldate, personid, receieverid, status
-  } = request.body;
+  if (request.userType == 2 || request.userType == 1) {
+    const {
+      pickaddressid, dropaddressid, pickdate, arrivaldate, personid, receieverid, status
+    } = request.body;
 
-  pool.query('INSERT INTO "Orders" ("PickAddressID", "DropAddressID", "PickDate", "ArrivalDate", "PersonID","ReceiverPersonID", "Status") VALUES ($1, $2, $3, $4, $5, $6, $7) RETURNING *',
-    [pickaddressid, dropaddressid, pickdate, arrivaldate, personid, receieverid, status], (error, result) => {
-      if (error) {
-        throw error;
-      }
-      response.status(201).json(result.rows[0]);
-    });
+    pool.query('INSERT INTO "Orders" ("PickAddressID", "DropAddressID", "PickDate", "ArrivalDate", "PersonID","ReceiverPersonID", "Status") VALUES ($1, $2, $3, $4, $5, $6, $7) RETURNING *',
+      [pickaddressid, dropaddressid, pickdate, arrivaldate, personid, receieverid, status], (error, result) => {
+        if (error) {
+          throw error;
+        }
+        response.status(201).json(result.rows[0]);
+      });
   }
-  else{
-    response.status(401).send("You cannnot perform this operation");    
+  else {
+    response.status(401).send("You cannnot perform this operation");
   }
 };
 
@@ -131,22 +128,27 @@ const createOrder = (request, response) => {
  * @param {obj} response response object * 
  */
 const updateOrder = (request, response) => {
-  const id = parseInt(request.params.id, 10);
-  const {
-    PickAddressID, DropAddressID, PickDate, ArrivalDate, PersonID, ReceiverPersonID, Status
-  } = request.body;
+  if (request.userType == 2 || request.userType == 1) {
+    const id = parseInt(request.params.id, 10);
+    const {
+      PickAddressID, DropAddressID, PickDate, ArrivalDate, PersonID, ReceiverPersonID, Status
+    } = request.body;
 
-  pool.query(
-    'UPDATE "Orders" SET "PickAddressID" = $1, "DropAddressID" = $2, "PickDate" = $3, "ArrivalDate" = $4, "PersonID"=$5,'+
-    ' "ReceiverPersonID"=$6, "Status"=$7 WHERE "OrderID" = $8',
-    [PickAddressID, DropAddressID, PickDate, ArrivalDate, PersonID, ReceiverPersonID, Status, id],
-    (error, results) => {
-      if (error) {
-        throw error;
-      }
-      response.sendStatus(200);
-    },
-  );
+    pool.query(
+      'UPDATE "Orders" SET "PickAddressID" = $1, "DropAddressID" = $2, "PickDate" = $3, "ArrivalDate" = $4, "PersonID"=$5,' +
+      ' "ReceiverPersonID"=$6, "Status"=$7 WHERE "OrderID" = $8',
+      [PickAddressID, DropAddressID, PickDate, ArrivalDate, PersonID, ReceiverPersonID, Status, id],
+      (error, results) => {
+        if (error) {
+          throw error;
+        }
+        response.sendStatus(200);
+      },
+    );
+  }
+  else {
+    response.status(401).send("You cannnot perform this operation");
+  }
 };
 
 /**
@@ -156,14 +158,18 @@ const updateOrder = (request, response) => {
  * @param {obj} response response object * 
  */
 const deleteOrder = (request, response) => {
-  const id = parseInt(request.params.id, 10);
+  if (request.userType == 2 || request.userType == 1) {
+    const id = parseInt(request.params.id, 10);
 
-  pool.query('DELETE FROM "Orders" WHERE "OrderID" = $1', [id], (error, results) => {
-    if (error) {
-      throw error;
-    }
-    response.sendStatus(200);
-  });
+    pool.query('DELETE FROM "Orders" WHERE "OrderID" = $1', [id], (error, results) => {
+      if (error) {
+        throw error;
+      }
+      response.sendStatus(200);
+    });
+  } else {
+    response.status(401).send("You cannnot perform this operation");
+  }
 };
 
 
@@ -176,22 +182,22 @@ const deleteOrder = (request, response) => {
 const getOrdersByUser = (request, response) => {
   const userId = parseInt(request.params.userid, 10);
 
-  pool.query('SELECT 	"Orders".*, "Person".*,	"addrop"."StreetAddress" as dropstreetAddress, 	"addrop"."City" as dropcity,'+
-	'"addrop"."Country" as dropcountry, "addrop"."PostCode" as droppostcode, 	"adpick"."StreetAddress" as pickstreetAddress,'+
-	'"adpick"."City" as pickcity,	"adpick"."Country" as pickcountry,	"adpick"."PostCode" as pickpostcode FROM "Orders" inner join "Address" addrop on "Orders"."DropAddressID"="addrop"."AddressID" '+
-  'inner join "Address" adpick on "Orders"."PickAddressID"="adpick"."AddressID" inner join "Person" on "Orders"."ReceiverPersonID" = "Person"."ID"'+
-  'where "Orders"."PersonID"= $1', [userId], (error, results) => {
-    if (error) {
-      throw error;
-    }
-    response.status(200).json(results.rows);
-  });
+  pool.query('SELECT 	"Orders".*, "Person".*,	"addrop"."StreetAddress" as dropstreetAddress, 	"addrop"."City" as dropcity,' +
+    '"addrop"."Country" as dropcountry, "addrop"."PostCode" as droppostcode, 	"adpick"."StreetAddress" as pickstreetAddress,' +
+    '"adpick"."City" as pickcity,	"adpick"."Country" as pickcountry,	"adpick"."PostCode" as pickpostcode FROM "Orders" inner join "Address" addrop on "Orders"."DropAddressID"="addrop"."AddressID" ' +
+    'inner join "Address" adpick on "Orders"."PickAddressID"="adpick"."AddressID" inner join "Person" on "Orders"."ReceiverPersonID" = "Person"."ID"' +
+    'where "Orders"."PersonID"= $1', [userId], (error, results) => {
+      if (error) {
+        throw error;
+      }
+      response.status(200).json(results.rows);
+    });
 };
 
 
 //private method to authenticate based on user type and company type
-const authenticate = (request, response, query) =>{
-  
+const authenticate = (request, response, query) => {
+
   if (request.userType == 2) {
     query += ' Where "CompanyId"=' + request.PersonRole;
